requirements.txt
torch>=2.0.0
transformers>=4.37.0
gradio>=4.0.0
Pillow>=10.0.0
accelerate>=0.25.0
qwen-vl-utils
